Update troubleshooting-efs-mounting.md
Added section for error message that happens when the EFS file system does not have the specified path.  Note that the AWS Create Access Point page is flat-out wrong when it says that it will create that folder for you.  I just spent half a day troubleshooting this issue, until I finally thought to look at the EFS file system contents despite the promise that the folder would be there.